Declare storage as using "maps" not Objects
The `Object` type has become stricter in Flow and means an Object with a
defined set of keys. WebStorage uses JS Objects as "maps," and Flow's
syntax for that use is `{ [key: string]: any}`. This offers slightly
more correct types and ensures calls to `get`, `set`, etc. can be
correctly chosen by ~0.90.0+ versions of Flow.